lush: fix 2x2 shuffled fader with non-DIR_UP
2x2 shuffled fader used an internal representation of DIR_UP
which then was incorrectly mapped to the pixel representation
which was non-DIR_UP when expanded.  This caused out of bound memory
accesses.

Solve the problem by requiring patterns always use an explicit
x,y location when drawing pixels and translate between an internal
state (likely still in DIR_UP) explicitly.